requirements.txt
python-dotenv
pyttsx3
SpeechRecognition
openai
screen-brightness-control
pillow
pyaudio
pypiwin32
pycaw
pyautogui
comtypes
